genai/api/smart-npc/src/requirements.txt
# Copyright 2024 Google LLC All Rights Reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
asyncpg==0.30.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloud-sql-python-connector==1.5.0
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
docstring_parser==0.16
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.115.5
filetype==1.2.0
frozenlist==1.5.0
gcloud==0.18.3
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.59.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.3.3
google-cloud-logging==3.11.3
google-cloud-resource-manager==1.13.1
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.62.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
jsonpath-python==1.0.6
langdetect==1.0.9
lxml==5.3.0
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
olefile==0.47
opentelemetry-api==1.28.2
packaging==24.2
pandas==2.2.3
pg8000==1.30.3
pgvector==0.2.3
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyparsing==3.2.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
RapidFuzz==3.10.1
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scramp==1.4.5
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.7
starlette==0.41.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.9
unstructured-client==0.28.1
urllib3==2.2.3
uvicorn==0.32.1
webencodings==0.5.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
